requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
tensorflow>=2.12.0
pillow>=10.0.0
numpy>=1.23.0
requests>=2.31.0

